+++ src/chilldkg/mod.rs
@@ -1,0 +1,4 @@
+pub mod chilldkg;
+pub mod dkg_process;
+mod coordinator;
+mod participant;
