+++ src/frost/mod.rs
@@ -1,4 +1,2 @@
 pub mod frost;
-pub mod chilldkg;
-pub mod dkg_process;
 pub mod signing_process;
